+++ packages/editor/src/models/Selectable.ts
@@ -582,47 +582,6 @@ export class Selectable {
   }
 
   static fromJSON(project: Project, json: SelectableJSON): Selectable {
-    if (json.original?.type === "component") {
-      // create instance
-
-      const node = project.nodes.create("instance");
-      node.name = json.name;
-      const selectable = node.selectable;
-      // TODO: position
-      selectable.style.mainComponent = json.original.id;
-
-      return selectable;
-    }
-
-    if (json.original?.type === "variant") {
-      throw new Error("TODO: pasting variant");
-    }
-
-    if (json.original?.type === "instance") {
-      const mainComponent = json.style.mainComponent;
-      if (mainComponent && project.nodes.get(mainComponent)) {
-        // original component exists in the project
-
-        const node = project.nodes.create("instance");
-        node.name = json.name;
-        const selectable = node.selectable;
-
-        const loadOverride = (json: SelectableJSON) => {
-          const idPath = json.id.split(":");
-          idPath[0] = node.id;
-          const selectable = project.selectables.get(idPath);
-          selectable.selfStyle.loadJSON(json.selfStyle ?? {});
-
-          for (const child of json.children) {
-            loadOverride(child);
-          }
-        };
-        loadOverride(json);
-
-        return selectable;
-      }
-    }
-
     const node = project.nodes.create(json.type);
     node.name = json.name;
     const selectable = node.selectable;
